test: add tests for escrow contract module closes #718
